# docs/openapi.yaml
openapi: 3.0.2
info:
  title: Prime Studio API
  version: 0.1.0
paths:
...
  /tag:
    put:
      summary: Put Tag
      description: Put tag to DynamoDB
      operationId: put_tag_tag_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReqTagPut"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    post:
      summary: Post Tag
      description: Post tag to DynamoDB
      operationId: post_tag_tag_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReqTagPost"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    delete:
      summary: Delete Tag
      description: delete tag meta and video/tag relations
      operationId: delete_tag_tag_delete
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReqTagDelete"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /category:
    put:
      summary: Put Category
      description: Put category to DynamoDB
      operationId: put_category_category_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReqCategoryPut"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    post:
      summary: Post Category
      description: Post category to DynamoDB
      operationId: post_category_category_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReqCategoryPost"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /videos/{category_id}:
    get:
      summary: Get Videos Contain Any Category
      description: get videos contain any tag
      operationId: get_videos_contain_any_category_videos__category_id__get
      parameters:
        - required: true
          schema:
            title: Category Id
          name: category_id
          in: path
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /category/{category_id}:
    delete:
      summary: Delete Category
      description: delete category meta
      operationId: delete_category_category__category_id__delete
      parameters:
        - required: true
          schema:
            title: Category Id
          name: category_id
          in: path
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
components:
  schemas:
...
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
...
    ReqCategoryPost:
      title: ReqCategoryPost
      required:
        - user
        - name
        - description
      type: object
      properties:
        user:
          title: User
          type: string
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
        note:
          title: Note
          type: string
          default: ""
        parentId:
          title: Parentid
          type: string
          default: C999
      description: Post category request body
    ReqCategoryPut:
      title: ReqCategoryPut
      required:
        - PK
        - user
      type: object
      properties:
        PK:
          title: Pk
          type: string
        user:
          title: User
          type: string
        name:
          title: Name
          type: string
        parentId:
          title: Parentid
          type: string
        description:
          title: Description
          type: string
        note:
          title: Note
          type: string
        invalid:
          title: Invalid
          type: boolean
      description: Put category request body
...
    ReqTagDelete:
      title: ReqTagDelete
      required:
        - PK
        - user
      type: object
      properties:
        PK:
          title: Pk
          type: string
        user:
          title: User
          type: string
      description: Delete tag request body
    ReqTagPost:
      title: ReqTagPost
      required:
        - user
        - name
      type: object
      properties:
        user:
          title: User
          type: string
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
          default: ""
        note:
          title: Note
          type: string
          default: ""
      description: Post tag request body
    ReqTagPut:
      title: ReqTagPut
      required:
        - PK
        - user
      type: object
      properties:
        PK:
          title: Pk
          type: string
        user:
          title: User
          type: string
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
        note:
          title: Note
          type: string
        invalid:
          title: Invalid
          type: boolean
      description: Put tag request body
...
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            type: string
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
...